refactor: Add rootReducer. WIP.
