--- src/store/city-slice/index.ts
@@ -1,2 +1,0 @@
-import { citySlice } from './city-slice';
-export default citySlice;
