--- src/store/offer-slice/index.ts
@@ -1,2 +1,0 @@
-import { offerSlice } from './offer-slice';
-export default offerSlice;
