--- src/store/review-slice/index.ts
@@ -1,2 +1,0 @@
-import { reviewSlice } from './review-slice';
-export default reviewSlice;
