+++ src/store/selectors.ts
@@ -1,4 +1,4 @@
-import { AuthorizationStatus } from '../const/api-const';
+import { AuthorizationStatus } from '../const';
 import { State } from './reducer';
 
 const selectAuthorizationStatus = (state: State) => state.authorizationStatus;
