+++ src/store/slices/offers-slice.ts
@@ -1,39 +1,37 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { NameSpace } from '../../const/store-const';
-import { OfferSlice } from '../../types/store-types';
+import { OffersSlice } from '../../types/store-types';
 import { RequestStatus } from '../../const/api-const';
+import { CITIES } from '../../const/app-const';
+import { City } from '../../types/app-types';
 import { SortOption } from '../../components/sort/const';
 import { SortOptionType } from '../../components/sort/types';
 import {
   getOfferPreviews,
-  getOfferFull,
-  getNearOfferPreviews,
   getFavoriteOffers,
 } from '../api-actions';
 
-const initialState: OfferSlice = {
+const initialState: OffersSlice = {
+  city: CITIES.Paris,
   sortOption: SortOption[0],
 
   offerPreviews: [],
   offerPreviewsStatus: RequestStatus.Idle,
 
-  offerFull: null,
-  offerFullStatus: RequestStatus.Idle,
-
-  nearOfferPreviews: [],
-  nearOfferPreviewsStatus: RequestStatus.Idle,
-
   favoriteOfferPreviews: [],
   favoriteOfferPreviewsStatus: RequestStatus.Idle,
 };
 
-export const offerSlice = createSlice({
-  name: NameSpace.Offer,
+export const offersSlice = createSlice({
+  name: NameSpace.Offers,
   initialState,
   reducers: {
+    setCity(state, action: PayloadAction<City>) {
+      state.city = action.payload;
+    },
     setSortOption(state, action: PayloadAction<SortOptionType>) {
       state.sortOption = action.payload;
-    }
+    },
   },
   extraReducers(builder) {
     builder
@@ -47,26 +45,6 @@ export const offerSlice = createSlice({
       .addCase(getOfferPreviews.rejected, (state) => {
         state.offerPreviewsStatus = RequestStatus.Failed;
       })
-      .addCase(getOfferFull.pending, (state) => {
-        state.offerFullStatus = RequestStatus.Loading;
-      })
-      .addCase(getOfferFull.fulfilled, (state, action) => {
-        state.offerFull = action.payload;
-        state.offerFullStatus = RequestStatus.Success;
-      })
-      .addCase(getOfferFull.rejected, (state) => {
-        state.offerFullStatus = RequestStatus.Failed;
-      })
-      .addCase(getNearOfferPreviews.pending, (state) => {
-        state.nearOfferPreviewsStatus = RequestStatus.Loading;
-      })
-      .addCase(getNearOfferPreviews.fulfilled, (state, action) => {
-        state.nearOfferPreviews = action.payload.slice(0, 3);
-        state.nearOfferPreviewsStatus = RequestStatus.Success;
-      })
-      .addCase(getNearOfferPreviews.rejected, (state) => {
-        state.nearOfferPreviewsStatus = RequestStatus.Failed;
-      })
       .addCase(getFavoriteOffers.pending, (state) => {
         state.favoriteOfferPreviewsStatus = RequestStatus.Loading;
       })
@@ -79,3 +57,5 @@ export const offerSlice = createSlice({
       });
   },
 });
+
+export const offersReducer = offersSlice.reducer;
