--- src/store/user-slice/index.ts
@@ -1,2 +1,0 @@
-import { userSlice } from './user-slice';
-export default userSlice;
